V1.2/INTEL-SE/requirements.txt
requests==2.31.0
PyYAML==6.0.1
python-docx==1.1.2
transformers==4.44.2
faiss-cpu==1.8.0
torch==2.4.1
reportlab==4.2.2
numpy==1.26.4
pdfplumber==0.11.4
